Added perissinotti model (Not validated)
This model fails to reproduce the results from the paper (Perissinotti 2018). The paper does not report a value for ai* for model 2 so it is assumed to be 0 (default starting values are reported in VGC-KiMo but I could not find the fitted rates). Model 1 appears to match reasonably for 1a but not well for 1b